chore: NO-JIRA refactor ButtonDocs component
