+++ src/docs-components/ButtonDocs/LoadingButton.tsx
@@ -1,0 +1,14 @@
+import { useLoading } from '../hooks/useLoading';
+
+import { Button, ButtonProps } from '@/components/Button';
+
+export const LoadingButton = (props: ButtonProps) => {
+  const [isLoading, startLoading] = useLoading();
+  return (
+    <Button
+      {...props}
+      state={isLoading ? 'loading' : undefined}
+      onClick={startLoading}
+    />
+  );
+};
