+++ src/docs-components/ButtonDocs/index.ts
@@ -1,0 +1,1 @@
+export { ButtonDocs } from './ButtonDocs';
